Moved info_alloc to radix-engine-profiling project
