+++ radix-engine-profiling/Cargo.toml
@@ -23,3 +23,4 @@ default = []
 resource_tracker = ["dep:shared_memory"]
 rocksdb = ["dep:plotters", "dep:linreg", "dep:blake2", "dep:rand", "radix-engine-stores/std", 
            "radix-engine-store-interface/std", "radix-engine-interface/std"]
+cpu_ram_metrics = []
