+++ radix-engine/Cargo.toml
@@ -19,7 +19,7 @@ lru = { version = "0.8.1", default-features = false, optional = true}
 moka = { version = "0.9.9", features = ["sync"], default-features = false, optional = true }
 strum = { version = "0.24", default-features = false, features = ["derive"] }
 perfcnt = { version = "0.8.0", optional = true }
-radix-engine-profiling = { path = "../radix-engine-profiling", optional = true, features = ["resource_tracker"] }
+radix-engine-profiling = { path = "../radix-engine-profiling", optional = true }
 resources-tracker-macro = { path = "../radix-engine-profiling/resources-tracker-macro" }
 paste = { version = "1.0.13" }
 radix-engine-macros = { path = "../radix-engine-macros", default-features = false }
@@ -64,7 +64,7 @@ alloc = ["sbor/alloc", "native-sdk/alloc", "transaction/alloc", "radix-engine-in
 
 # Enables heap memory and CPU cycles resource tracing - available only for Linux OS on x86 arch.
 # Requires CAP_PERFMON capability for the process (sudo setcap cap_perfmon=eip <exec_file>).
-cpu_ram_metrics = ["std", "dep:perfcnt"]
+cpu_ram_metrics = ["std", "dep:perfcnt",  "dep:radix-engine-profiling", "radix-engine-profiling?/cpu_ram_metrics"]
 
 # Use `wasmer` as WASM engine, otherwise `wasmi`
 wasmer = ["dep:wasmer", "dep:wasmer-compiler-singlepass"]
@@ -74,7 +74,7 @@ wasmer = ["dep:wasmer", "dep:wasmer-compiler-singlepass"]
 moka = ["dep:moka"]
 lru = ["dep:lru"]
 
-resource_tracker = ["dep:radix-engine-profiling", "resources-tracker-macro/resource_tracker", "radix-engine-common/resource_tracker"]
+resource_tracker = ["dep:radix-engine-profiling", "radix-engine-profiling?/resource_tracker", "resources-tracker-macro/resource_tracker", "radix-engine-common/resource_tracker"]
 
 # System/Application Database Checker
 db_checker = []
